requirements.txt
agno==1.7.6
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
beautifulsoup4==4.13.4
brotli==1.1.0
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.3
curl-cffi==0.12.0
debugpy==1.8.15
decorator==5.2.1
deprecation==2.1.0
distro==1.9.0
docstring-parser==0.17.0
duckduckgo-search==8.1.1
ecdsa==0.19.1
et-xmlfile==2.0.0
executing==2.2.0
fastapi==0.116.1
ffmpy==0.6.1
filelock==3.18.0
frozendict==2.4.6
fsspec==2025.7.0
gitdb==4.0.12
gitpython==3.1.45
google-auth==2.40.3
google-genai==1.27.0
gradio==5.38.2
gradio-client==1.11.0
greenlet==3.2.3
groovy==0.1.2
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.2
idna==3.10
ipykernel==6.30.0
ipython==9.4.0
ipython-pygments-lexers==1.1.1
ipywidgets==8.1.7
jedi==0.19.2
jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jupyter-client==8.6.3
jupyter-core==5.8.1
jupyterlab-widgets==3.0.15
lancedb==0.24.2
langchain==0.3.27
langchain-core==0.3.72
langchain-text-splitters==0.3.9
langsmith==0.4.8
lxml==6.0.0
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multitasking==0.0.12
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.2
openai==1.97.1
openpyxl==3.1.5
orjson==3.11.1
overrides==7.7.0
packaging==25.0
pandas==2.3.1
parso==0.8.4
peewee==3.18.2
pillow==11.3.0
platformdirs==4.3.8
primp==0.15.0
prompt-toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
pure-eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pydub==0.25.1
pygments==2.19.2
pylance==0.32.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-http-client==3.3.7
python-multipart==0.0.20
pytz==2025.2
pywin32==311
pyyaml==6.0.2
pyzmq==27.0.0
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.1.0
rsa==4.9.1
ruff==0.12.5
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.1
semantic-version==2.10.0
sendgrid==6.12.4
sentence-transformers==5.0.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
sqlalchemy==2.0.41
stack-data==0.6.3
starlette==0.47.2
sympy==1.14.0
tantivy==0.24.0
tenacity==8.5.0
threadpoolctl==3.6.0
tokenizers==0.21.4
tomli==2.2.1
tomlkit==0.13.3
torch==2.7.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.54.0
typer==0.16.0
typing-extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.13
websockets==15.0.1
werkzeug==3.1.3
widgetsnbextension==4.0.14
yfinance==0.2.65
zstandard==0.23.0
